requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
pydantic==2.9.2
python-dotenv>=1.0.1
openai>=1.0.0
httpx==0.27.2
python-docx>=1.1.2
reportlab>=4.2.2